Fix 64 bit platform
git-svn-id: http://root.cern.ch/svn/root/trunk@12160 27541ba8-7e3a-0410-8455-c3a389f83636

+++ cint/src/bc_exec.cxx
@@ -35,7 +35,7 @@ extern "C" int G__bc_exec_virtualbase_bytecode(G__value *result7
 
   // TODO, below is a dead copy of virtual_bytecode which may not work
 
-  int vtagnum = (int)funcname; // tagnum of given pointer
+  long vtagnum = (long)funcname; // tagnum of given pointer
   int vtblindex = hash&0xffff; // virtual function table index
   int vbasetagnum = hash/0x10000; // vbasetagnum
   int voffset=G__struct.virtual_offset[vtagnum]; // offset for true tagnum info
@@ -100,7 +100,7 @@ extern "C" int G__bc_exec_virtual_bytecode(G__value *result7
 			,int hash              // vtblindex,basetagnum
 			) {
 
-  int vtagnum = (int)funcname; // tagnum of given pointer
+  long vtagnum = (long)funcname; // tagnum of given pointer
   int vtblindex = hash&0xffff; // virtual function table index
   int vbasetagnum = hash/0x10000; // vbasetagnum
   int voffset=G__struct.virtual_offset[vtagnum]; // offset for true tagnum info
